Criação do CRUD de Mesa e Categoria Finalizado.

+++ src/main/java/br/senac/sp/poo/lanchonete/model/Categoria.java
@@ -2,11 +2,8 @@ package br.senac.sp.poo.lanchonete.model;
 
 import jakarta.persistence.*;
 import lombok.Data;
-import lombok.Generated;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.util.List;
 
 @Data
 @Getter
@@ -15,7 +12,8 @@ import java.util.List;
 public class Categoria {
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
+
     private String nome;
     private String descricao;
 }
